task 2/EmotionRecognition/requirements.txt
# Core dependencies
python>=3.8
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
scipy>=1.7.0
tqdm>=4.62.0

# Audio processing
librosa>=0.9.0
soundfile>=0.10.0
pydub>=0.25.0

# Deep learning frameworks
tensorflow>=2.8.0
# Alternative: uncomment if using PyTorch
# torch>=1.10.0
# torchaudio>=0.10.0

# Web and API (optional)
flask>=2.0.0
flask-cors>=3.0.0

# Development and testing
jupyter>=1.0.0
ipython>=7.0.0
black>=22.0.0
pytest>=6.0.0
pytest-cov>=3.0.0

# Data handling
h5py>=3.6.0
pyyaml>=6.0.0

# Audio recording (for real-time prediction)
pyaudio>=0.2.11
